requirements.txt
# Core dependencies
tensorflow>=2.13.0
tensorflow-recommenders>=0.7.0
numpy>=1.23.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Data processing
scipy>=1.10.0

# System monitoring
psutil>=5.9.0

# Optional: Fast retrieval
faiss-cpu>=1.7.4  # Use faiss-gpu for GPU support

# Optional: Experiment tracking
wandb>=0.15.0

# Development tools
jupyter>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0

